move get_case_properties to dbaccessors
