+++ corehq/apps/importer/util.py
@@ -1,33 +1,20 @@
 import json
-import xlrd
 from collections import defaultdict
 from datetime import date
+
+import xlrd
 from django.utils.translation import ugettext_lazy as _
 from couchdbkit import NoResultFound
-from dimagi.utils.couch.database import get_db
+from xlrd import xldate_as_tuple
+
 from corehq.apps.importer.const import LookupErrors, ImportErrors
 from casexml.apps.case.models import CommCareCase
-from xlrd import xldate_as_tuple
 from corehq.apps.groups.models import Group
 from corehq.apps.users.cases import get_wrapped_owner
 from corehq.apps.users.models import CouchUser
 from corehq.apps.users.util import format_username
 from corehq.apps.locations.models import SQLLocation
 
-
-def get_case_properties(domain, case_type=None):
-    """
-    For a given case type and domain, get all unique existing case properties,
-    known and unknown
-    """
-    key = [domain]
-    if case_type:
-        key.append(case_type)
-    rows = get_db().view('hqcase/all_case_properties',
-                         startkey=key,
-                         endkey=key + [{}],
-                         reduce=True, group=True, group_level=3).all()
-    return sorted(set([r['key'][2] for r in rows]))
 
 class ImporterConfig(object):
     """
